requirements.txt
flask>=2.0.0
python-jenkins==1.8.0
python-dotenv>=0.19.0
setuptools>=65.5.1
importlib-metadata>=4.0.0
gunicorn>=20.1.0